Fix .csv scraping, df clean up
Some of the found urls get a raw text respnose, so we need to sort by url suffix (i.e. '.zip' | '.csv') to either unzip or just write to file.

The df was in a weird order, so we convert the 'VALIDATION_DATE' column using pd.to_datetime() and sort by the new date format, newest to oldest.